vehicle view implementation, including view model
